requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
torch-geometric>=2.4.0

# Vision & depth
transformers>=4.38.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.26.0

# Node label embedding (Step 1 → Step 2 adapter)
sentence-transformers>=2.7.0

# Scene graph utilities
scipy>=1.11.0
shapely>=2.0.0
matplotlib>=3.8.0
networkx>=3.2.0
tqdm>=4.66.0

# Config
PyYAML>=6.0
